New graph manager.
